backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-genai>=0.2.0
google-adk>=1.0.0
python-multipart>=0.0.7
supabase>=2.3.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pyjwt>=2.8.0
gotrue>=2.0.0